refactor(build): move html building and deleting out to separate function
